Be a little more consistent about use-package :init and :config
It's too much of a pain to figure out exactly the _best_ way to use
them; startup is probably faster if as much as possible is delayed until
the package is lazily loaded (i.e. using :config), but some things might
have to be configured beforehand (i.e. using :init) to work properly.

The current scheme generally uses :init to set variables/properties and
:config to call functions or enable modes.