Initial emacs lsp changes
- Add use-package declaration for lsp-mode with basic config. It's not
enabled on the after-init-hook, but rather invoked manually, because I
don't want it most of the time.
- Add company-mode hook for lsp-mode, which toggles company-mode state
when lsp-mode state is updated. lsp-mode automatically enables
company-mode if it isn't already enabled, but disabling lsp-mode leaves
company-mode on. This is probably not the ideal behavior, because I
would generally want both to be enabled or disabled together.
company-mode can still be toggled independently.